修改了文件结构，针对不同模块设计了不同的请求参数
