+++ src/components/course/CourseManagerHomePage.vue
