+++ src/router/index.js
@@ -1,13 +1,14 @@
 import {createRouter, createWebHistory,onBeforeRouteLeave } from 'vue-router';
 
-const rolehome =['teacherhomne','adminhome','superadminhome','secretariatehome','academicaffairshome','coursemanagerhome',
-    'deanhome','departmenthome','majormanagerhome'];
+const rolehome =['teacherhomne','adminhome','superadminhome','secretariatehome','academicaffairshome','deanhome',
+    'departmenthome','majormanagerhome','coursemanagerhome',];
 // superadminhome 超级管理员首页
 // secretariatehome 教学秘书首页
 // academicaffairshome 教务处首页
 // deanhome 学院负责人首页
 // departmenthome 系主任首页
-// majormanagerhome 专业负责人首页  
+// majormanagerhome 专业负责人首页 
+// coursemanagerhome 课程负责人首页
 
 const routes = [
     {
@@ -28,82 +29,82 @@ const routes = [
             {
                 path: 'sysmangt/rolemangt',
                 name: 'Rolemangt', //角色管理
-                component: () => import('../components/Rolemangt.vue')
+                component: () => import('../components/admin/Rolemangt.vue')
             },
             {
                 path: 'sysmangt/rolepurview',
                 name: 'Rolepurview', //角色授权
-                component: () => import('../components/Rolepurview.vue')
+                component: () => import('../components/admin/Rolepurview.vue')
             },
             {
                 path: 'sysmangt/termmangt',
                 name: 'TermsManagement', //学期管理
-                component: () => import('../components/TermMangt.vue')
+                component: () => import('../components/admin/TermMangt.vue')
             },
             {
                 path: 'sysmangt/creatteachunit',
                 name: 'Unitssmangt', //教学单位管理
-                component: () => import('../components/CreateTeachUnit.vue')
+                component: () => import('../components/admin/CreateTeachUnit.vue')
             },
             {
                 path: 'sysmangt/peoplemangt',
                 name: 'PeopleManagement', //人员管理
-                component: () => import('../components/Peoplemangt.vue')
+                component: () => import('../components/admin/Peoplemangt.vue')
             },
             {
                 path: 'sysmangt/collegemangt',
                 name: 'Collegecollegemangt', //学院管理
-                component: () => import('../components/Collegemangt.vue')
+                component: () => import('../components/admin/Collegemangt.vue')
             },
             {
                 path: 'sysmangt/departmentmangt',
                 name: 'Departmentmangt', //部门管理
-                component: () => import('../components/Departmentmangt.vue')
+                component: () => import('../components/admin/Departmentmangt.vue')
             },
             {
                 path: 'sysmangt/professionmangt',
                 name: 'Professionmangt', //专业管理
-                component: () => import('../components/Professionmangt.vue')
+                component: () => import('../components/admin/Professionmangt.vue')
             },
             {
                 path: 'sysmangt/classmangt',
                 name: 'Classmangt', //专业班级管理
-                component: () => import('../components/Classmangt.vue')
+                component: () => import('../components/admin/Classmangt.vue')
             },
             {
                 path: 'coursemangt/coursemangt',    // 课程管理
                 name: 'CourseManagement',
-                component: () => import('../components/Coursemangt.vue')
+                component: () => import('../components/course/Coursemangt.vue')
             },
             {
                 path: 'coursemangt/abilitydictionary',  // 能力字典
                 name: 'AbilityDictionary',
-                component: () => import('../components/AbilityDictionary.vue')
+                component: () => import('../components/course/AbilityDictionary.vue')
             },
             {
                 path: 'coursemanagerhomepage',  // 课程负责人首页
                 name: 'CourseManagerHomePage',
-                component: () => import('../components/CourseManagerHomePage.vue')
+                component: () => import('../components/course/CourseManagerHomePage.vue')
             },
             {
                 path: 'evasys/courseinfo/formative/keyword',    // 形成性评价模型-关键字
                 name: 'Keyword',
-                component: () => import('../components/Keyword.vue')
+                component: () => import('../components/evaluation/Keyword.vue')
             },
             {
                 path: 'evasys/courseinfo/formative/ability',    // 形成性评价模型-能力
                 name: 'Ability',
-                component: () => import('../components/Ability.vue')
+                component: () => import('../components/evaluation/Ability.vue')
             },
             {
                 path: 'evasys/courseinfo/formative/teachingobjectives', // 形成性评价模型-基本教学目标
                 name: 'TeachingObjectives',
-                component: () => import('../components/TeachingObjectives.vue')
+                component: () => import('../components/evaluation/TeachingObjectives.vue')
             },
             {
                 path: 'evasys/courseinfo/formative/knowledgeunit', // 形成性评价模型-知识单元
                 name: 'KnowledgeUnit',
-                component: () => import('../components/KnowledgeUnit.vue')
+                component: () => import('../components/evaluation/KnowledgeUnit.vue')
             },
         ]
     },
