+++ src/views/HomePage.vue
@@ -235,17 +235,17 @@ onMounted(() => {
     sessionStorage.removeItem('users');
     sessionStorage.removeItem('isLoggedIn');
 
-    router.push({ name: 'Login' });
+    // router.push({ name: 'Login' });
     // 或
     // ElMessage.error('请重新登录');
   }
   //获取完pinia中的数据后重新重定向到父页面
-  router.push(homeurl.value);
+  // router.push(homeurl.value);
   // 0304：为生成侧面导航栏此处暂时写死：当前接口为：POST /homes/superadminhome
-// request.post(`${homeurl}`,loginInfo)
+// request.admin.post(`${homeurl}`,loginInfo)
 // 0311修改为teacherhome
 //获取菜单栏的数据
-  request.post(`/homes/teacherhome`,toRaw(loginInfo))
+  request.admin.post(`/homes/teacherhome`,toRaw(loginInfo))
       .then(res => {
         // 登录成功
         if (res.code === 200 && res.data.length > 0)  {
